+++ pgpainless-core/src/main/java/org/pgpainless/key/generation/type/RSA.java
@@ -16,18 +16,39 @@
 package org.pgpainless.key.generation.type;
 
 import javax.annotation.Nonnull;
+import java.security.spec.AlgorithmParameterSpec;
+import java.security.spec.RSAKeyGenParameterSpec;
 
 import org.pgpainless.algorithm.PublicKeyAlgorithm;
 import org.pgpainless.key.generation.type.length.RsaLength;
 
-public class RSA_SIGN extends RSA_GENERAL {
+/**
+ * Key type that specifies the RSA_GENERAL algorithm.
+ */
+public class RSA implements KeyType {
 
-    RSA_SIGN(@Nonnull RsaLength length) {
-        super(length);
+    private final RsaLength length;
+
+    RSA(@Nonnull RsaLength length) {
+        this.length = length;
+    }
+
+    public static RSA withLength(@Nonnull RsaLength length) {
+        return new RSA(length);
+    }
+
+    @Override
+    public String getName() {
+        return "RSA";
     }
 
     @Override
     public PublicKeyAlgorithm getAlgorithm() {
-        return PublicKeyAlgorithm.RSA_SIGN;
+        return PublicKeyAlgorithm.RSA_GENERAL;
+    }
+
+    @Override
+    public AlgorithmParameterSpec getAlgorithmSpec() {
+        return new RSAKeyGenParameterSpec(length.getLength(), RSAKeyGenParameterSpec.F4);
     }
 }
