--- pgpainless-core/src/main/java/org/pgpainless/key/generation/type/RSA_ENCRYPT.java
@@ -1,33 +1,0 @@
-/*
- * Copyright 2018 Paul Schaub.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.pgpainless.key.generation.type;
-
-import javax.annotation.Nonnull;
-
-import org.pgpainless.algorithm.PublicKeyAlgorithm;
-import org.pgpainless.key.generation.type.length.RsaLength;
-
-public class RSA_ENCRYPT extends RSA_GENERAL {
-
-    RSA_ENCRYPT(@Nonnull RsaLength length) {
-        super(length);
-    }
-
-    @Override
-    public PublicKeyAlgorithm getAlgorithm() {
-        return PublicKeyAlgorithm.RSA_ENCRYPT;
-    }
-}
